Rewrite, drop headless chrome in favor of userscripts.
Removes dependency on headless_chrome in favor of
userscripts. This way it's system agnostic and
you're no longer at the mercy of headless_chrome
not bugging out.

Userscripts located in their own folder, and
are pretty much a copy of the existing observer
(youtube only so far), that send to the
websocket directly.

Just open a livestream and enable the userscript,
it should work in normal and pop-out mode.

Note: default port changed to 3000, just becuase
it's easier to remember.